# api/swagger.yaml
openapi: 3.0.1
info:
  title: Swagger API
  version: 1.0.0
servers:
  - url: 'http://localhost:3000/'
paths:
...
  /note:
    post:
      tags:
        - Note
      summary: Create note 
      operationId: createNote
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
                answer:
                  type: string
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Parent category didn't found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /note/{id}:
    get:
      tags:
        - Note
      summary: Find note by ID
      operationId: getNoteById
      parameters:
        - name: id
          in: path
          description: ID of note that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'


        '404':
          description: Note didn't found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422':
          description: Invalid id parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags:
        - Note
      summary: Create note 
      operationId: createNoteInParentCategory
      parameters:
        - name: id
          in: path
          description: ID of parent category 
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
                answer:
                  type: string
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Parent category didn't found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422':
          description: Invalid id parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      tags:
        - Note
      summary: update note 
      operationId: updateNote
      parameters:
        - name: id
          in: path
          description: ID of note 
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
                answer:
                  type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
        '404':
          description: Note did't found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422':
          description: Invalid id parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags:
        - Note
      summary: delete note 
      operationId: deleteNote
      parameters:
        - name: id
          in: path
          description: ID of note 
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
        '404':
          description: Note did't found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422':
          description: Invalid id parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
components:
  schemas:
    Note:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        answer:
          type: string
        parentId:
          type: string
        authorId:
          type: string
        createdAt:
          type: string
          format: date
        updatedAt:
          type: string
          format: date
      xml:
        name: Note
...
    ApiResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer
          format: int32
        data:
          type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT